SAS-48: Render user's cases on home page (#2)
* Add new client and service for cases endpoints

* Add assertions for important logs

But suppress excessive error logging in relevant tests :)

* Improve integration test setup

This commit simplifies the command to run the integration tests, so that
the wiremock setup is done automatically, and the server is correctly
awaited, before starting the tests.

* Render response from API on home page

For now, the home page is set as the user's list of assigned cases.

* Match API response and add CRN

* Remove example service

* Use generated API types